Added todo list, some minor changes

+++ software/rp6/func_protos.h
@@ -55,10 +55,11 @@ struct SD{
 	uint8_t bumperRight:1;
 	uint8_t bumperLeft:1;
 	int16_t ultrasonic;
-	int16_t motorLeft;
-	int16_t motorRight;
+	int8_t motorLeft;
+	int8_t motorRight;
 	uint8_t ledStatus; //8 all leds in one variable
 };
+
 
 extern struct SD sensorData;
